pipeline/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pipeline/requirements.in -o pipeline/requirements.txt
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.7
    # via requests
colorama==0.4.6
    # via tqdm
cython==0.29.37
    # via hdbscan
feedparser==6.0.11
    # via -r pipeline/requirements.in
filelock==3.28.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2026.3.0
    # via
    #   huggingface-hub
    #   torch
hdbscan==0.8.33
    # via -r pipeline/requirements.in
huggingface-hub==0.36.2
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via requests
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via
    #   hdbscan
    #   pynndescent
    #   scikit-learn
llvmlite==0.47.0
    # via
    #   numba
    #   pynndescent
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numba==0.65.0
    # via
    #   pynndescent
    #   umap-learn
numpy==1.26.4
    # via
    #   -r pipeline/requirements.in
    #   hdbscan
    #   numba
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
    #   umap-learn
packaging==26.1
    # via
    #   huggingface-hub
    #   transformers
pillow==12.2.0
    # via sentence-transformers
pynndescent==0.6.0
    # via umap-learn
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
regex==2026.4.4
    # via transformers
requests==2.33.1
    # via
    #   -r pipeline/requirements.in
    #   huggingface-hub
    #   transformers
safetensors==0.7.0
    # via transformers
scikit-learn==1.4.2
    # via
    #   -r pipeline/requirements.in
    #   hdbscan
    #   pynndescent
    #   sentence-transformers
    #   umap-learn
scipy==1.17.1
    # via
    #   hdbscan
    #   pynndescent
    #   scikit-learn
    #   sentence-transformers
    #   umap-learn
sentence-transformers==2.7.0
    # via -r pipeline/requirements.in
setuptools==81.0.0
    # via torch
sgmllib3k==1.0.0
    # via feedparser
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
torch==2.11.0
    # via sentence-transformers
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
    #   umap-learn
transformers==4.57.6
    # via sentence-transformers
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   torch
umap-learn==0.5.6
    # via -r pipeline/requirements.in
urllib3==2.6.3
    # via requests
